+++ annofabcli/statistics/table.py
@@ -6,7 +6,6 @@ from typing import Any, Dict, List, Optional
 
 import dateutil
 import more_itertools
-import numpy
 import pandas
 from annofabapi.dataclass.statistics import ProjectAccountStatisticsHistory, WorktimeStatistics, WorktimeStatisticsItem
 from annofabapi.models import InputDataId, Inspection, InspectionStatus, Task, TaskHistory, TaskPhase, TaskStatus
@@ -953,57 +952,6 @@ class Table:
             phase_list.remove(TaskPhase.ACCEPTANCE.value)
 
         return phase_list
-
-    @staticmethod
-    def merge_productivity_per_user_from_aw_time(df1: pandas.DataFrame, df2: pandas.DataFrame) -> pandas.DataFrame:
-        """
-        ユーザごとの生産性・品質情報が格納されたDataFrameを結合する
-        Args:
-            df1:
-            df2:
-
-        Returns:
-            マージ済のユーザごとの生産性・品質情報
-        """
-
-        def max_last_working_date(date1, date2):
-            if not isinstance(date1, str) and numpy.isnan(date1):
-                date1 = ""
-            if not isinstance(date2, str) and numpy.isnan(date2):
-                date2 = ""
-            max_date = max(date1, date2)
-            if max_date == "":
-                return numpy.nan
-            else:
-                return max_date
-
-        def merge_row(row1: pandas.Series, row2: pandas.Series) -> pandas.Series:
-            string_column_list = ["username", "biography", "last_working_date"]
-            sum_row = row1.drop(labels=string_column_list, level=0).fillna(0) + row2.drop(
-                labels=string_column_list, level=0
-            ).fillna(0)
-            sum_row.loc["username", ""] = row1.loc["username", ""]
-            sum_row.loc["biography", ""] = row1.loc["biography", ""]
-            sum_row.loc["last_working_date", ""] = max_last_working_date(
-                row1.loc["last_working_date", ""], row2.loc["last_working_date", ""]
-            )
-            return sum_row
-
-        user_id_set = set(df1["user_id"]) | set(df2["user_id"])
-        sum_df = df1.set_index("user_id").copy()
-        added_df = df2.set_index("user_id")
-
-        for user_id in user_id_set:
-            if user_id not in added_df.index:
-                continue
-            if user_id in sum_df.index:
-                sum_df.loc[user_id] = merge_row(sum_df.loc[user_id], added_df.loc[user_id])
-            else:
-                sum_df.loc[user_id] = added_df.loc[user_id]
-
-        phase_list = Table._get_phase_list(list(sum_df["monitored_worktime_hour"].columns))
-        _add_ratio_column_for_productivity_per_user(sum_df, phase_list=phase_list)
-        return sum_df.reset_index().sort_values(["user_id"])
 
     def create_labor_df(self, df_labor: Optional[pandas.DataFrame] = None) -> pandas.DataFrame:
         """
